MNT: rename private module to clarify that it's not part of public API
