+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 
 import cmyt  # noqa: F401
-from cmyt.utils import create_cmap_overview
+from cmyt._utils import create_cmap_overview
 
 mpl_compare = pytest.mark.mpl_image_compare(
     savefig_kwargs={"bbox_inches": "tight"},
